==> Code/3-model.ipynb <==
import os
os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
# Seed value
seed_value = 55

# 1. Set the `PYTHONHASHSEED` environment variable at a fixed value
import os
os.environ['PYTHONHASHSEED']=str(seed_value)

# 2. Set the `python` built-in pseudo-random generator at a fixed value
import random
random.seed(seed_value)

# 3. Set the `numpy` pseudo-random generator at a fixed value
import numpy as np
np.random.seed(seed_value)

# 4. Set the `tensorflow` pseudo-random generator at a fixed value
import tensorflow as tf
tf.random.set_seed(seed_value)

session_conf = tf.compat.v1.ConfigProto(intra_op_parallelism_threads=1, inter_op_parallelism_threads=1)
sess = tf.compat.v1.Session(graph=tf.compat.v1.get_default_graph(), config=session_conf)
tf.compat.v1.keras.backend.set_session(sess)

import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
%matplotlib inline
from tensorflow import keras
from sklearn.model_selection import train_test_split

dataset_file_name = 'human_dataset.txt'
y_train_full = np.load('../Saved Data/Training Set/labels.npy')
X_train_full = np.load('../Saved Data/Training Set/dna_spectral_representation.npy')

k = 5
input_vector_size = np.power(4, k)

X_train, X_valid, y_train, y_valid = train_test_split(X_train_full, y_train_full, test_size=0.1,
                                                        random_state=20)

# Model : LeNet5 like Architecture CNN for Text Classification

model = keras.Sequential()

model.add(keras.layers.Conv1D(filters=30, kernel_size=(5), activation='relu', input_shape=(input_vector_size, 1),padding="same"))
model.add(keras.layers.MaxPooling1D())
model.add(keras.layers.Dropout(0.3))

model.add(keras.layers.Conv1D(filters=15, kernel_size=(5), activation='relu', input_shape=(int(input_vector_size / 2), 1), padding="same"))
model.add(keras.layers.MaxPooling1D())
model.add(keras.layers.Dropout(0.3))

model.add(keras.layers.Flatten())

model.add(keras.layers.Dense(units=256, activation='relu'))
model.add(keras.layers.Dropout(0.3))

model.add(keras.layers.Dense(units=128, activation='relu'))
model.add(keras.layers.Dropout(0.3))

model.add(keras.layers.Dense(units=64, activation='relu'))

model.add(keras.layers.Dense(units=int(7), activation = 'softmax'))

model.summary()

opt = keras.optimizers.SGD(learning_rate=0.045, momentum=0.0)
model.compile(loss="sparse_categorical_crossentropy", optimizer=opt, metrics=["accuracy"])

history = model.fit(X_train, y_train, epochs=30, validation_data=(X_valid, y_valid), batch_size=32)

# data = history.history
# pd.DataFrame(data).plot(figsize=(8, 6))
# plt.grid(True)
# plt.gca().set_ylim(0, 2)
# plt.show()

model.save("../Saved Data/Model/")
model.save("../Saved Data/Model Architecture/model.h5")
